+++ packages/terra-form-select/src/_Frame.jsx
@@ -12,6 +12,10 @@ const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
+   * Text for the clear option.
+   */
+  clearOptionDisplay: PropTypes.string,
+  /**
    * Whether the select is disabled.
    */
   disabled: PropTypes.bool,
@@ -36,6 +40,11 @@ const propTypes = {
    * The max height of the dropdown.
    */
   maxHeight: PropTypes.number,
+  /**
+   * The maximum number of options that can be selected. A value less than 2 will be ignored.
+   * Only applicable to variants allowing multiple selections (e.g.; `multiple`; `tag`).
+   */
+  maxSelectionCount: PropTypes.number,
   /**
    * Content to display when no search results are found.
    */
@@ -85,10 +94,12 @@ const propTypes = {
 };
 
 const defaultProps = {
+  clearOptionDisplay: undefined,
   disabled: false,
   dropdown: undefined,
   dropdownAttrs: undefined,
   isInvalid: false,
+  maxSelectionCount: undefined,
   noResultContent: undefined,
   onDeselect: undefined,
   onSearch: undefined,
@@ -355,12 +366,14 @@ class Frame extends React.Component {
 
   render() {
     const {
+      clearOptionDisplay,
       disabled,
       display,
       dropdown,
       dropdownAttrs,
       isInvalid,
       maxHeight,
+      maxSelectionCount,
       noResultContent,
       onDeselect,
       onSearch,
@@ -431,6 +444,8 @@ class Frame extends React.Component {
                  onSelect: this.handleSelect,
                  onRequestClose: this.closeDropdown,
                  searchValue: this.state.searchValue,
+                 maxSelectionCount,
+                 clearOptionDisplay,
                })}
           </Dropdown>
           )