+++ packages/terra-form-select/src/_Menu.jsx
@@ -6,6 +6,8 @@ import 'terra-base/lib/baseStyles';
 import KeyCode from 'keycode-js';
 import Variants from './_constants';
 import AddOption from './_AddOption';
+import ClearOption from './_ClearOption';
+import MaxSelection from './_MaxSelection';
 import NoResults from './_NoResults';
 import Util from './_MenuUtil';
 import styles from './_Menu.module.scss';
@@ -19,6 +21,15 @@ const propTypes = {
    * The content of the menu.
    */
   children: PropTypes.node,
+  /**
+   * Text for the clear option.
+   */
+  clearOptionDisplay: PropTypes.string,
+  /**
+   * The maximum number of options that can be selected. A value less than 2 will be ignored.
+   * Only applicable to variants allowing multiple selections (e.g.; `multiple`; `tag`).
+   */
+  maxSelectionCount: PropTypes.number,
   /**
    * Content to display when no results are found.
    */
@@ -61,6 +72,8 @@ const propTypes = {
 
 const defaultProps = {
   children: undefined,
+  clearOptionDisplay: undefined,
+  maxSelectionCount: undefined,
   noResultContent: undefined,
   onDeselect: undefined,
   optionFilter: undefined,
@@ -100,26 +113,40 @@ class Menu extends React.Component {
    */
   static getDerivedStateFromProps(props, state) {
     const {
-      searchValue, noResultContent,
+      clearOptionDisplay, maxSelectionCount, searchValue, noResultContent,
     } = props;
-    const children = Util.filter(props.children, props.searchValue, props.optionFilter);
 
+    let children;
     let hasNoResults = false;
+    let hasMaxSelection = false;
+    let hasAddOption = false;
 
-    if (Util.shouldAllowFreeText(props, children)) {
-      children.push(<AddOption value={searchValue} />);
-    }
-
-    if (Util.shouldShowNoResults(props, children)) {
-      children.push(<NoResults noResultContent={noResultContent} value={searchValue} />);
-      hasNoResults = true;
+    if (searchValue && searchValue.length > 0 && Util.isMaxSelectionReached(props)) {
+      children = [(<MaxSelection value={maxSelectionCount} />)];
+      hasMaxSelection = true;
     } else {
-      hasNoResults = false;
+      children = Util.filter(props.children, props.searchValue, props.optionFilter);
+      children = Util.updateSelectionState(children, props);
+
+      if (Util.shouldAllowFreeText(props, children)) {
+        children.push(<AddOption value={searchValue} />);
+        hasAddOption = true;
+      }
+
+      if (Util.shouldShowNoResults(props, children)) {
+        children.push(<NoResults noResultContent={noResultContent} value={searchValue} />);
+        hasNoResults = true;
+      }
+
+      if (Util.shouldShowClearOption(props, hasAddOption, hasNoResults)) {
+        children.unshift(<ClearOption display={clearOptionDisplay} value="" />);
+      }
     }
 
     return {
       children,
       searchValue,
+      hasMaxSelection,
       hasNoResults,
       active: Util.getActiveOptionFromProps(props, children, state),
     };
@@ -147,6 +174,7 @@ class Menu extends React.Component {
 
     this.liveRegionTimeOut = setTimeout(() => {
       const {
+        hasMaxSelection,
         hasNoResults,
       } = this.state;
 
@@ -166,6 +194,8 @@ class Menu extends React.Component {
 
       if (hasNoResults && searchValue) {
         visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.noResults' }, { text: searchValue });
+      } else if (hasMaxSelection) {
+        visuallyHiddenComponent.current.innerText = intl.formatMessage({ id: 'Terra.form.select.maxSelectionOption' }, { text: this.props.maxSelectionCount });
       } else {
         visuallyHiddenComponent.current.innerText = '';
       }