+++ packages/terra-form-select/CHANGELOG.md
@@ -6,6 +6,12 @@ Unreleased
 ### Changed
 * Replaced local keyCode values with keycode-js npm package
 
+5.9.0 - (March 6th, 2019)
+------------------
+### Added
+* Added option to limit the maximum number of selection count for multi-select variants.
+* Introduce an option to allow clearing the currently selected item
+
 5.8.0 - (March 5th, 2019)
 ------------------
 ### Fixed